move internal preview actions
